+++ tests/testthat/test-omics-feature_merge.R
@@ -1,3 +1,19 @@
+## Load example data
+taxa <- metagenomics$new(
+  biomData = "input/metagenomics/biom_with_taxonomy_hdf5.biom",
+  metaData = "input/metagenomics/metadata.tsv",
+  treeData = "input/metagenomics/rooted_tree.newick"
+)
+
+test_that("`omics$feature_merge()` -- Argument checks", {
+
+})
+
+test_that("`omics$feature_merge()` -- Behavioral checks", { 
+ 
+})
+
+
 test_that("Testing `feature_merge`", {
   taxa <- metagenomics$new(
     biomData = "input/metagenomics/biom_with_taxonomy_hdf5.biom",